an absolute ton of documentation/typespec
