+++ app.config.js
@@ -1,26 +1,25 @@
 require('dotenv').config();
 
-const appConfig = {
-    // config vars
-    get ENV(){ return process.env.NODE_ENV || 'local'; },
-    get PORT(){ return process.env.PORT || 8080; },
-    get DOMAIN(){ return process.env.domain || `http://localhost:${this.PORT}` },
-    get DISCORD_BOT_TOKEN(){return process.env.discord_bot_token},
-    get DISCORD_BOT_NAME(){return 'little-daiko'},
+class AppConfig{
+    get ENV(){ return process.env.NODE_ENV || 'local'; };
+    get PORT(){ return process.env.PORT || 8080; };
+    get DOMAIN(){ return process.env.domain || `http://localhost:${this.PORT}` };
+    get DISCORD_BOT_TOKEN(){return process.env.discord_bot_token};
+    get DISCORD_BOT_NAME(){return 'little-daiko'};
 
-    get BOT(){ return require('./src/bot')},
+    get MILDOM_CLIENT(){ return require('./src/adapters/mildom/mildom.client')};
+    get DISCORD_CLIENT(){ return require('./src/adapters/discord/discord.client')};
+    get DISCORD_HELPERS(){ return require('./src/adapters/discord/discord.helpers')};
 
-    get MILDOM_CLIENT(){ return require('./src/adapters/mildom/mildom.client')},
-    get DISCORD_CLIENT(){ return require('./src/adapters/discord/discord.client')},
-    get DISCORD_HELPERS(){ return require('./src/adapters/discord/discord.helpers')},
+    get CONFIG_STORAGE(){ return require('./src/persistence/config.storage')};
+    get LISTENER_STORAGE(){ return require('./src/persistence/listener.storage')};
 
-    get CONFIG_STORAGE(){ return require('./src/persistence/config.storage')},
-    get LISTENER_STORAGE(){ return require('./src/persistence/listener.storage')},
-
-    get EVENTS(){ return require('./src/models/event')},
-    get COMMANDS(){ return require('./src/models/command')},
-    get PERMISSIONS(){ return require('./src/models/permissions')},
-    // get LOGGER(){ return require('./src/utils/logger')},
+    get BOT(){ return require('./src/engine/bot')};
+    get EVENTS(){ return require('./src/engine/event')};
+    get COMMANDS(){ return require('./src/engine/commands')};
+    get PERMISSIONS(){ return require('./src/engine/permissions')};
 }
 
-module.exports.AppConfig = appConfig;
+const instance = new AppConfig();
+
+module.exports.AppConfig = instance;
