+++ index.js
@@ -1,7 +1,7 @@
 const { AppConfig } = require('./app.config');
 
 async function main(){
-    const bot = await AppConfig.BOT.initialize(AppConfig, 'system');
+    const bot = await AppConfig.BOT.startBot(AppConfig, 'system');
     bot.login(AppConfig.DISCORD_BOT_TOKEN);
 }
 main();
