+++ src/engine/event.js
@@ -1,7 +1,26 @@
+const { AppConfig } = require("../../app.config");
+const { Message } = require('discord.js');
+
+/**
+ * The callback for an event
+ *
+ * @callback EventCallback
+ * @param {Message} message
+ * @param {ErrorCallback} onError
+ */
+
+/**
+ * The callback for an error
+ *
+ * @callback ErrorCallback
+ * @param {Message} message
+ * @param {*} Exception
+ */
+
 /**
  * List of events for the discord client to listen for
- * @param {AppConfig} appConfig 
- * @returns List of events
+ * @param {AppConfig} appConfig The dependency injection config 
+ * @returns {Map<string, EventCallback>} Map of events indexed by name
  */
 function events(appConfig) { 
     const discordHelpers = appConfig.DISCORD_HELPERS;
@@ -22,7 +41,7 @@ function events(appConfig) {
             const role = appConfig.CONFIG_STORAGE.getProperty(message, 'role');
             for(let entry of appConfig.COMMANDS(appConfig)){
                 if(entry.aliases.includes(command)){
-                    if(appConfig.PERMISSIONS(appConfig)[String(entry.permissions)].check(message, role.ops)){
+                    if(appConfig.PERMISSIONS(appConfig).get(entry.permissionLevel).check(message, role.ops)){
                         entry.callback(message, args).then((out) => { 
                             if(out){
                                 message.react(out);
@@ -30,10 +49,10 @@ function events(appConfig) {
                         }).catch((e) => { 
                             onError(message, e);
                         });
-                        return;
                     }else{
                         message.channel.send('You do not have permission to use that command.');
                     }
+                    return;
                 }
             }
         }]
